aoneportfolio-backend/requirements.txt
Django
django
djangorestframework